requirements.txt
fastapi
uvicorn
python-multipart
pydantic
streamlit
requests
pillow
pytesseract
pandas
python-dotenv
